chore(automations): consolidate filter imports via the _filters package root
