+++ wandb/automations/_filters/run_metrics.py
@@ -16,8 +16,8 @@ from typing_extensions import override
 from wandb._pydantic import GQLBase
 from wandb.automations._validators import LenientStrEnum
 
-from .expressions import FilterExpr
-from .operators import BaseOp, RichReprResult
+from . import BaseOp, FilterExpr
+from .operators import RichReprResult
 
 if TYPE_CHECKING:
     from wandb.automations.events import RunMetricFilter
